breaking typeo into sections
